requirements.txt
-i https://pypi.org/simple
blinker==1.6.2 ; python_version >= '3.7'
click==8.1.3 ; python_version >= '3.7'
et-xmlfile==1.1.0 ; python_version >= '3.6'
flask==2.3.2
flask-cors==3.0.10
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
openpyxl==3.1.2
pillow==9.5.0
reportlab==4.0.4
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
werkzeug==2.3.6 ; python_version >= '3.8'
